+++ src/NEFDashboard/containers/setPassword/index.jsx
@@ -1,14 +1,43 @@
-import React from "react";
+import React, { useContext, useEffect, useState } from "react";
 import "./setPassword.css";
 import { Flex, Form, Input, Button, Divider } from "antd";
 import SetPasswordIcon from "../../../assets/newDashboardIcons/setPasswordIcon.svg";
 import { useNavigate } from "react-router-dom";
+import { MessageContext } from "../../../context/message";
+import backendAPI from "../../../api/backendAPI";
+import { useTranslation } from "react-i18next";
 
 const SetPasswordForm = () => {
-  const navigate = useNavigate();
+  const { setInfoMessage, setErrorMessage } = useContext(MessageContext);
 
-  const onFinish = (values) => {
-    console.log("Success:", values);
+  const backend_API = new backendAPI();
+  const { t } = useTranslation();
+  const navigate = useNavigate();
+  const [token, setToken] = useState(null);
+
+  useEffect(() => {
+    const urlParams = new URLSearchParams(window.location.search);
+    if (urlParams.has("token")) {
+      const paramValue = urlParams.get("token");
+      setToken(paramValue);
+    }
+  }, []);
+
+  const onFinish = async (data) => {
+    try {
+      const response = await backend_API.resetPassword(
+        data?.newPassword,
+        token,
+      );
+      if (response && response?.newPassword) {
+        setErrorMessage(response?.newPassword);
+        return;
+      }
+      setInfoMessage(t("messages.success.passwordReset"));
+      navigate("/login");
+    } catch (error) {
+      setErrorMessage(t("messages.error.updatePassword"));
+    }
   };
   const onFinishFailed = (errorInfo) => {
     console.log("Failed:", errorInfo);
@@ -46,7 +75,8 @@ const SetPasswordForm = () => {
             rules={[
               {
                 required: true,
-                message: "Please input your new Password!",
+                message: "Password must be at least 8 characters!",
+                min: 8,
               },
             ]}
           >
@@ -59,8 +89,19 @@ const SetPasswordForm = () => {
             rules={[
               {
                 required: true,
-                message: "Please input your confirm Password!",
+                message: "Password must be at least 8 characters!",
+                min: 8,
               },
+              ({ getFieldValue }) => ({
+                validator(_, value) {
+                  if (!value || getFieldValue("newPassword") === value) {
+                    return Promise.resolve();
+                  }
+                  return Promise.reject(
+                    new Error("Password and Confirm Password must be match!"),
+                  );
+                },
+              }),
             ]}
           >
             <Input.Password placeholder="Confirm your new password" />